requirements.txt
Django==1.7.3
GDAL==1.11.1
GitPython==0.3.6
argcomplete==0.8.4
click==4.0
django-cors-headers==1.0.0
django-debug-toolbar==1.2.2
django-rq==0.7.0
djangorestframework==3.0.3
geopy==1.9.1
gitdb==0.6.4
pytesseract==0.1.5
redis==2.10.3
requests==2.5.1
rq==0.5.1
smmap==0.9.0
sqlparse==0.1.14
webops-django==0.1
wsgiref==0.1.2
xlrd==0.9.3
